+++ index.html
@@ -44,6 +44,6 @@
 </head>
 <body>
     <div id="root">Not Rendered</div>
-    <script type="module" src="./App.js"></script>
+    <script type="module" src="src/App.js"></script>
 </body>
 </html>
