+++ notes/.~lock.episode5_lets_get_hooked.odt#
@@ -1,0 +1,1 @@
+,shanmukha,shanmukha-Latitude-3420,23.10.2024 21:59,file:///home/shanmukha/.config/libreoffice/4;
