+++ src/utils/constants.js
@@ -1,0 +1,2 @@
+export const HEADER_IMAGE = "https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcQwv45iAbWWwD5glterGTBIO6Q6pnjukcFkGw&usqp=CAU"
+export const CLOUDINARY_IMAGE= "https://media-assets.swiggy.com/swiggy/image/upload/fl_lossy,f_auto,q_auto,w_660/"
